Rename pkg/plugin/provider -> pkg/provider
