+++ pkg/provider/cleanup.go
@@ -1,4 +1,4 @@
-package pkg
+package provider
 
 import (
 	"fmt"
