+++ pkg/provider/codegen/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"os"
 
-	"github.com/aiyengar2/Rancher-Plugin-gMSA/pkg/plugin/provider/codegen/generator"
+	"github.com/aiyengar2/Rancher-Plugin-gMSA/pkg/provider/codegen/generator"
 	v1 "k8s.io/api/core/v1"
 )
